Keerthana.G.S-3122225001059-CSE-B/requirements.txt
requests>=2.31.0
pymongo>=4.8.0
python-dotenv>=1.0.1
urllib3>=2.2.0
